rename viewsets and serializers
