+++ backend/core/serializers/organizations.py
@@ -1,7 +1,8 @@
-from core.models import Organization
-from core.serializers.users_serializers import GetUserSerializer
 from rest_framework import serializers
 from rest_framework.validators import UniqueValidator
+
+from core.models import Organization
+from core.serializers.users_serializers import GetUserSerializer
 
 
 class CreateOrganizationSerializer(serializers.ModelSerializer):
