+++ backend/core/urls.py
@@ -1,8 +1,9 @@
-from core.views.comments_viewset import CommentsViewSet
-from core.views.organizations_viewset import OrganizationsViewSet
-from core.views.tickets_viewset import TicketsViewSet
-from core.views.users_viewset import UsersViewSet
 from django.urls import path
+
+from core.views.comments import CommentsViewSet
+from core.views.organizations import OrganizationsViewSet
+from core.views.tickets import TicketsViewSet
+from core.views.users import UsersViewSet
 
 user_list = UsersViewSet.as_view({"get": "list", "post": "create"})
