+++ backend/core/views/comments.py
@@ -1,42 +1,42 @@
 import logging
 
 from django.db import DatabaseError, IntegrityError
-from django.db.models import ObjectDoesNotExist, Prefetch
+from django.db.models import ObjectDoesNotExist, Prefetch, Q
 from rest_framework import status, viewsets
 from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 
 from core.models import Comment, Organization, Ticket, User
-from core.serializers.organizations_serializers import (
-    CreateOrganizationSerializer,
-    GetOrganizationSerializer,
-    PartialUpdateOrganizationSerializer,
-    UpdateOrganizationSerializer,
+from core.serializers.comments import (
+    CreateCommentSerializer,
+    GetCommentSerializer,
+    PartialUpdateCommentSerializer,
+    UpdateCommentSerializer,
 )
 
 logger = logging.getLogger(__name__)
 
 
-class OrganizationsViewSet(viewsets.ViewSet):
+class CommentsViewSet(viewsets.ViewSet):
     def list(self, request):
-        queryset = Organization.objects.prefetch_related("members").all()
-        serializer = GetOrganizationSerializer(queryset, many=True)
+        queryset = Comment.objects.select_related("author").all()
+        serializer = GetCommentSerializer(queryset, many=True)
 
         return Response(serializer.data)
 
     def retrieve(self, request, pk=None):
         try:
-            organization = Organization.objects.prefetch_related("members").get(pk=pk)
+            comment = Comment.objects.select_related("author").get(pk=pk)
         except ObjectDoesNotExist:
             return Response(
-                {"error": "Organization not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "Comment not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
-        serializer = GetOrganizationSerializer(organization)
+        serializer = GetCommentSerializer(comment)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def create(self, request):
-        serializer = CreateOrganizationSerializer(data=request.data)
+        serializer = CreateCommentSerializer(data=request.data)
 
         try:
             serializer.is_valid(raise_exception=True)
@@ -46,13 +46,40 @@ class OrganizationsViewSet(viewsets.ViewSet):
             return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
 
         try:
-            organization = Organization.objects.create(
-                name=validated_data["name"],
-                email=validated_data["email"],
-            )
+            author_id = validated_data.get("author_id")
+            ticket_id = validated_data.get("ticket_id")
 
-            organization.save()
-            organization.objects.prefetch_related("members")
+            if not author_id:
+                return Response(
+                    {"error": "Author id should be provided"},
+                    status=status.HTTP_404_NOT_FOUND,
+                )
+
+            try:
+                user = User.objects.get(id=author_id)
+
+            except User.DoesNotExist:
+                return Response(
+                    {"error": "Author not found"}, status=status.HTTP_404_NOT_FOUND
+                )
+
+            try:
+                ticket = Ticket.objects.get(
+                    Q(id=ticket_id) & (Q(requestor=user) | Q(assignee=user))
+                )
+
+            except Ticket.DoesNotExist:
+                return Response(
+                    {
+                        "error": "Ticket not found or comment's author not assignee or requestor"
+                    },
+                    status=status.HTTP_403_FORBIDDEN,
+                )
+
+            comment = Comment.objects.create(
+                text=validated_data["text"], author=user, ticket=ticket
+            )
+            comment.save()
 
         except ValidationError as e:
             return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
@@ -71,23 +98,23 @@ class OrganizationsViewSet(viewsets.ViewSet):
 
         except Exception as e:
             return Response(
-                {"error": f"Failed to create organization: {str(e)}"},
+                {"error": f"Failed to create comment: {str(e)}"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
-        response = GetOrganizationSerializer(organization)
+        response = GetCommentSerializer(comment)
         return Response(response.data, status=status.HTTP_201_CREATED)
 
     def update(self, request, pk=None):
         try:
-            organization = Organization.objects.prefetch_related("members").get(pk=pk)
+            comment = Comment.objects.select_related("author").get(pk=pk)
 
         except ObjectDoesNotExist:
             return Response(
-                {"error": "Organization not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "Comment not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
-        serializer = UpdateOrganizationSerializer(data=request.data, partial=False)
+        serializer = UpdateCommentSerializer(data=request.data, partial=False)
 
         try:
             serializer.is_valid(raise_exception=True)
@@ -97,10 +124,9 @@ class OrganizationsViewSet(viewsets.ViewSet):
             return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
 
         try:
-            organization.name = validated_data.get("name", organization.name)
-            organization.email = validated_data.get("email", organization.email)
+            comment.text = validated_data.get("text", comment.text)
 
-            organization.save()
+            comment.save()
 
         except ValidationError as e:
             return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
@@ -119,34 +145,32 @@ class OrganizationsViewSet(viewsets.ViewSet):
 
         except Exception as e:
             return Response(
-                {"error": f"Failed to update organization: {str(e)}"},
+                {"error": f"Failed to update comment: {str(e)}"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
-        response = GetOrganizationSerializer(organization)
+        response = GetCommentSerializer(comment)
         return Response(response.data, status=status.HTTP_200_OK)
 
     def partial_update(self, request, pk=None):
         try:
-            organization = Organization.objects.prefetch_related("members").get(pk=pk)
+            comment = Comment.objects.select_related("author").get(pk=pk)
 
         except ObjectDoesNotExist:
             return Response(
-                {"error": "Organization not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "Comment not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
-        serializer = PartialUpdateOrganizationSerializer(
-            data=request.data, partial=True
-        )
+        serializer = PartialUpdateCommentSerializer(data=request.data, partial=True)
 
         try:
             serializer.is_valid(raise_exception=True)
             validated_data = serializer.validated_data
 
             for attr, value in validated_data.items():
-                setattr(organization, attr, value)
+                setattr(comment, attr, value)
 
-            organization.save()
+            comment.save()
 
         except ValidationError as e:
             return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
@@ -165,28 +189,28 @@ class OrganizationsViewSet(viewsets.ViewSet):
 
         except Exception as e:
             return Response(
-                {"error": f"Failed to update organization: {str(e)}"},
+                {"error": f"Failed to update comment: {str(e)}"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
-        response = GetOrganizationSerializer(organization)
+        response = GetCommentSerializer(comment)
         return Response(response.data, status=status.HTTP_200_OK)
 
     def destroy(self, request, pk=None):
         try:
-            organization = Organization.objects.get(pk=pk)
+            comment = Comment.objects.get(pk=pk)
 
-            organization.delete()
-            logger.info(f"Organization {organization.id} deleted successfully!")
+            comment.delete()
+            logger.info(f"Comment {comment.id} deleted successfully!")
 
             return Response(
-                {"status": "Organization deleted successfully!"},
+                {"status": "Comment deleted successfully!"},
                 status=status.HTTP_204_NO_CONTENT,
             )
 
         except ObjectDoesNotExist:
             return Response(
-                {"error": "Organization not found"}, status=status.HTTP_404_NOT_FOUND
+                {"error": "Comment not found"}, status=status.HTTP_404_NOT_FOUND
             )
 
         except DatabaseError as e:
@@ -197,6 +221,6 @@ class OrganizationsViewSet(viewsets.ViewSet):
 
         except Exception as e:
             return Response(
-                {"error": f"Failed to delete user {str(e)}"},
+                {"error": f"Failed to delete comment {str(e)}"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
