+++ backend/core/views/tickets.py
@@ -1,3 +1,11 @@
+from core.models import Comment, Organization, Ticket, User
+from core.serializers.tickets import (
+    CreateTicketSerializer,
+    GetTicketSerializer,
+    PartialUpdateTicketSerializer,
+    UpdateTicketSerializer,
+)
+
 import logging
 
 from django.db import DatabaseError, IntegrityError
@@ -5,14 +13,6 @@ from django.db.models import ObjectDoesNotExist, Prefetch, Q
 from rest_framework import status, viewsets
 from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
-
-from core.models import Comment, Organization, Ticket, User
-from core.serializers.tickets_serializers import (
-    CreateTicketSerializer,
-    GetTicketSerializer,
-    PartialUpdateTicketSerializer,
-    UpdateTicketSerializer,
-)
 
 logger = logging.getLogger(__name__)
