+++ backend/helpdesk/urls.py
@@ -1,7 +1,4 @@
 from django.contrib import admin
 from django.urls import include, path
 
-urlpatterns = [
-    path("admin/", admin.site.urls),
-    path("helpdesk/", include("core.urls"))
-    ]
+urlpatterns = [path("admin/", admin.site.urls), path("helpdesk/", include("core.urls"))]
